refactor(extension-common): メッセージング定義とフレーム追跡を移す
拡張機能内・ウィンドウ間のメッセージング定義 (activeTab / frameCas / overlay)、
ウィンドウメッセージングの基盤、フレーム準備状態の追跡を extension-common へ
移す。

位置算出と描画 (is-frame-visible、use-* フック、マーカーの .tsx) はビュー要件で
分岐しうるため、2 つのビューが揃うまで inspector に残す。
