+++ apps/inspector/src/components/frameCas/index.ts
@@ -1,6 +1,3 @@
-export * from "./extension-events";
 export * from "./is-frame-visible";
-export * from "./types";
 export * from "./use-frame-cas-location-consumer";
 export * from "./use-frame-cas-location-provider";
-export * from "./window-events";
