+++ apps/inspector/src/components/frameCas/use-frame-cas-location-provider.ts
@@ -1,6 +1,8 @@
-import { type FrameVerifiedCas } from "@originator-profile/extension-common";
+import {
+  type FrameVerifiedCas,
+  frameCasExtensionMessenger,
+} from "@originator-profile/extension-common";
 import { useEffect } from "react";
-import { frameCasExtensionMessenger } from "./extension-events";
 
 export function useFrameCasLocationProvider(
   tabId: number,
