+++ apps/inspector/src/components/frameCas/use-located-cas-coordinate.ts
@@ -1,8 +1,11 @@
+import {
+  FrameCasCoordinate,
+  FramesCasCoordinate,
+  frameCasWindowMessenger,
+} from "@originator-profile/extension-common";
 import { startTransition, useEffect, useState } from "react";
 import { useMap } from "react-use";
-import { FrameCasCoordinate, FramesCasCoordinate } from "./types";
 import { useFrameCasLocationConsumer } from "./use-frame-cas-location-consumer";
-import { frameCasWindowMessenger } from "./window-events";
 
 export function useLocatedCasCoordinate(): {
   framesCasCoordinate: FramesCasCoordinate;
