+++ apps/inspector/src/components/overlay/ContentsArea.tsx
@@ -1,8 +1,11 @@
+import {
+  CaCoordinate,
+  FrameCoordinate,
+} from "@originator-profile/extension-common";
 import { Target } from "@originator-profile/model";
 import { useId } from "react";
 import { useWindowSize } from "react-use";
 import { twMerge } from "tailwind-merge";
-import { CaCoordinate, FrameCoordinate } from "../frameCas/types";
 import { useFrameCaRects } from "../frameCas/use-frame-ca-rects";
 import { useLocatedCasCoordinate } from "../frameCas/use-located-cas-coordinate";
 import useElements from "./use-elements";
