+++ apps/inspector/src/components/overlay/index.ts
@@ -2,4 +2,3 @@ export * from "./CaMarker";
 export * from "./CasMap";
 export * from "./ContentsArea";
 export * from "./overlay";
-export * from "./window-events";
