--- apps/inspector/src/components/windowMessaging/index.ts
@@ -1,1 +1,0 @@
-export * from "./window-messaging";
