+++ packages/extension-common/src/index.ts
@@ -1,9 +1,17 @@
+export * from "./active-tab/events";
+export * from "./active-tab/frame-readiness-tracker";
 export * from "./credentials";
+export * from "./frame-cas/extension-events";
+export * from "./frame-cas/types";
+export * from "./frame-cas/window-events";
 export * from "./i18n";
 export * from "./link-verification";
+export * from "./overlay/extension-events";
+export * from "./overlay/window-events";
 export * from "./site-profile";
 export * from "./utils/navigation-state";
 export * from "./utils/persistent-map";
 export * from "./utils/problem-code";
 export * from "./utils/registry-ops";
 export * from "./utils/warning-params";
+export * from "./window-messaging";
