+++ packages/extension-common/src/window-messaging.ts
